requirements.txt
fastapi==0.115.2
uvicorn==0.30.1
numpy==1.26.4
torch==2.9.1
torchaudio==2.9.1
librosa==0.10.1
pydub==0.25.1
python-multipart==0.0.9
Pillow==10.2.0
scipy==1.11.4